Begin a refactor of the way you set up the scene
ContextManager is being replaced by Scene and Engine. Scene will contain objects
and cameras and such, and will not expose the GL stuff or need to know much about
the GLFW stuff. Engine will deal with most of the GL and GLFW setup and eventually
talk directly to the input on behalf of the scenes (via listeners unless I find
a better way to do something like that).

This was mostly motivated by ContextManager just kinda being the catchall for
all the window stuff. It has a bunch of static functions and attributes
that GLFW needs for callbacks, but passes itself off as an object that can
be instantiated (which it really shouldn't be). It also has to handle
initialization of things that shouldn't happen until a GL context has been
created. That stuff has been passed to Engine.

Currently, Engine is being set up as a singleton object, but that might
change depending on needs. Resize and mouse input need to be static
methods, so that might drive me away from the singleton. But I'd like
to avoid having essentially just member variables in a static class
(width and height for example).