Add support for key event callbacks in Engine and add back camera movement
The engine also gets the scene by reference when playing it rather than by value,
which was a bug.

Right now, callbacks are the easiest way I can think of to do key events,
but they're probably a little slower than optimal. They also only allow
for key presses, not key releases, and it's a polling system, so it will
call the callbacks as long as the last event was a press. There's a way to do
a callback for a single function that triggers every time a key changes state,
but you can only have 1 of those functions.