Add a mouse position event registration function and port the mouselook to the engine
This is an example of using a ``static'' callback for input events.
My inclination right now is to do polling for the mouse and keys
rather than callbacks for both, just because it means betraying
the singleton pattern a little less. I'd suspect there's a performance
loss when polling, but I need to research that.